Trait atium::algebra::expression::Expression

pub trait Expression: DynClone {
    // Required methods
    fn eval(&self) -> Box<dyn Expression>;
    fn simplify(&self) -> Box<dyn Expression>;
    fn as_any(&self) -> &dyn Any;
    fn debug(&self, indent: usize) -> String;
    fn to_typist(&self) -> String;
}

The Expression trait represents an algebraic expression. It provides methods for evaluating and simplifying the expression, as well as converting the expression to a debug string or a Typist string.

Required Methods

fn eval(&self) -> Box<dyn Expression>

Evaluates the expression and returns a new expression.

fn simplify(&self) -> Box<dyn Expression>

Simplifies the expression and returns a new simplified expression.

fn as_any(&self) -> &dyn Any

Returns a reference to the expression as a dyn Any, which can be downcast to its concrete type.

fn debug(&self, indent: usize) -> String

Returns a debug string for the expression. The indent parameter specifies the indentation level.

fn to_typist(&self) -> String

Returns a Typist string for the expression.